Fix stdin issue
